+++ examples/example_glfw_opengl2/Dielectric.h
@@ -4,9 +4,9 @@
 class Dielectric : public Material
 {
 public:
-    Dielectric(float ri); //n2/n1
+    Dielectric(double ri); //n2/n1
     virtual bool scatter(const Ray& r_in, const HitRecord& rec, vec3& attenuation, Ray& scattered) const;
 
-    float ref_idx;
+    double ref_idx;
 };
