+++ examples/example_glfw_opengl2/HitRecord.h
@@ -4,7 +4,7 @@ class Material;
 class HitRecord
 {
 public:
-    float t;     //命中射线的长度
+    double t;     //命中射线的长度
     vec3 p;      //命中终点坐标
     vec3 normal; //命中点的法向量
     Material* mat_ptr; //new
