+++ examples/example_glfw_opengl2/Ray.h
@@ -4,14 +4,14 @@ class Ray
 {
     vec3 A; //起点
     vec3 B; //方向
-    float tm; //时间
+    double tm; //时间
 
 public:
 
     Ray() {}
-    Ray(const vec3& a, const vec3& b,float t=0.0);
+    Ray(const vec3& a, const vec3& b,double t=0.0);
     vec3 origin() const;
     vec3 direction() const;
-    vec3 point_at_parameter(float t) const; //终点的坐标
-    float time() const;
+    vec3 point_at_parameter(double t) const; //终点的坐标
+    double time() const;
 };
