+++ examples/example_glfw_opengl2/Sphere.h
@@ -4,10 +4,10 @@ class Sphere : public Hittable
 {
 public:
     vec3 center;
-    float radius;
+    double radius;
     Material* mat_ptr; /* NEW */
     Sphere();
-    Sphere(vec3 cen, float r, Material* m); //new
+    Sphere(vec3 cen, double r, Material* m); //new
     //如果命中了，命中记录保存到rec
-    virtual bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const;
+    virtual bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const;
 };
