+++ examples/example_glfw_opengl2/vec3.h
@@ -15,36 +15,36 @@
 #include <stdlib.h>
 #include <iostream>
 
-static constexpr float M_PI = 3.1415926f;
+static constexpr double M_PI = 3.1415926f;
 
 class vec3  {
     public:
         vec3() {}
-        vec3(float e0, float e1, float e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
-        inline float x() const { return e[0]; }
-        inline float y() const { return e[1]; }
-        inline float z() const { return e[2]; }
-        inline float r() const { return e[0]; }
-        inline float g() const { return e[1]; }
-        inline float b() const { return e[2]; }
+        vec3(double e0, double e1, double e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
+        inline double x() const { return e[0]; }
+        inline double y() const { return e[1]; }
+        inline double z() const { return e[2]; }
+        inline double r() const { return e[0]; }
+        inline double g() const { return e[1]; }
+        inline double b() const { return e[2]; }
 
         inline const vec3& operator+() const { return *this; }
         inline vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-        inline float operator[](int i) const { return e[i]; }
-        inline float& operator[](int i) { return e[i]; }
+        inline double operator[](int i) const { return e[i]; }
+        inline double& operator[](int i) { return e[i]; }
 
         inline vec3& operator+=(const vec3 &v2);
         inline vec3& operator-=(const vec3 &v2);
         inline vec3& operator*=(const vec3 &v2);
         inline vec3& operator/=(const vec3 &v2);
-        inline vec3& operator*=(const float t);
-        inline vec3& operator/=(const float t);
+        inline vec3& operator*=(const double t);
+        inline vec3& operator/=(const double t);
 
-        inline float length() const { return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]); }
-        inline float squared_length() const { return e[0]*e[0] + e[1]*e[1] + e[2]*e[2]; }
+        inline double length() const { return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]); }
+        inline double squared_length() const { return e[0]*e[0] + e[1]*e[1] + e[2]*e[2]; }
         inline void make_unit_vector();
 
-        float e[3];
+        double e[3];
 };
 
 
@@ -59,7 +59,7 @@ inline std::ostream& operator<<(std::ostream &os, const vec3 &t) {
 }
 
 inline void vec3::make_unit_vector() {
-    float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+    double k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     e[0] *= k; e[1] *= k; e[2] *= k;
 }
 
@@ -79,19 +79,19 @@ inline vec3 operator/(const vec3 &v1, const vec3 &v2) {
     return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-inline vec3 operator*(float t, const vec3 &v) {
+inline vec3 operator*(double t, const vec3 &v) {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline vec3 operator/(vec3 v, float t) {
+inline vec3 operator/(vec3 v, double t) {
     return vec3(v.e[0]/t, v.e[1]/t, v.e[2]/t);
 }
 
-inline vec3 operator*(const vec3 &v, float t) {
+inline vec3 operator*(const vec3 &v, double t) {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline float dot(const vec3 &v1, const vec3 &v2) {
+inline double dot(const vec3 &v1, const vec3 &v2) {
     return v1.e[0] * v2.e[0]
          + v1.e[1] * v2.e[1]
          + v1.e[2] * v2.e[2];
@@ -131,15 +131,15 @@ inline vec3& vec3::operator-=(const vec3& v) {
     return *this;
 }
 
-inline vec3& vec3::operator*=(const float t) {
+inline vec3& vec3::operator*=(const double t) {
     e[0] *= t;
     e[1] *= t;
     e[2] *= t;
     return *this;
 }
 
-inline vec3& vec3::operator/=(const float t) {
-    float k = 1.0f/t;
+inline vec3& vec3::operator/=(const double t) {
+    double k = 1.0f/t;
 
     e[0] *= k;
     e[1] *= k;
@@ -158,10 +158,10 @@ inline vec3 reflect(const vec3& v, const vec3& n) {
 
 
 
-inline bool refract(const vec3& v, const vec3& n, float ni_over_nt, vec3& refracted) {
+inline bool refract(const vec3& v, const vec3& n, double ni_over_nt, vec3& refracted) {
     vec3 uv = unit_vector(v);
-    float dt = dot(uv, n);
-    float discriminant = 1.0 - ni_over_nt*ni_over_nt*(1-dt*dt);
+    double dt = dot(uv, n);
+    double discriminant = 1.0 - ni_over_nt*ni_over_nt*(1-dt*dt);
     if (discriminant > 0) {
         refracted = ni_over_nt*(uv - n*dt) - n*sqrt(discriminant);
         return true;
@@ -172,18 +172,18 @@ inline bool refract(const vec3& v, const vec3& n, float ni_over_nt, vec3& refrac
 
 
 
-inline float schlick(float cosine, float n1_over_n2) {
-    float r0 = (1-n1_over_n2) / (1+n1_over_n2);
+inline double schlick(double cosine, double n1_over_n2) {
+    double r0 = (1-n1_over_n2) / (1+n1_over_n2);
     r0 = r0*r0;
     return r0 + (1-r0)*pow((1 - cosine),5);
 }
 
-inline float random_float()
+inline double random_float()
 {
     return rand() / (RAND_MAX + 1.0);
 }
 
-inline float random_float(float a, float b)
+inline double random_float(double a, double b)
 {
     return (rand() / (RAND_MAX + 1.0)) * (b - a) + a;
 }
